backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.16.1
opencv-python==4.8.1.78
mediapipe==0.10.14
numpy==1.26.4
pillow==10.1.0
scikit-learn==1.3.2
pandas==2.1.4
bcrypt==4.1.1
pyjwt==2.8.0
python-dotenv==1.0.0
protobuf==4.25.3
